+++ harmony-api/Dockerfile
@@ -19,6 +19,10 @@ RUN if [ "$ENV" = "prod" ] ; then \
   cp ./environment/.env.prod ./2 ; \
   fi
 
+# Install Postgres Client
+RUN apk --no-cache add postgresql14-client
+ENTRYPOINT [ "psql" ]
+
 # Compile App
 RUN npx prisma generate --schema ./prisma/schema.prisma
 RUN npx tsc
