+++ harmony-api/package.json
@@ -14,6 +14,7 @@
     "posttest": "npm run lint",
     "connectprisma": "npx prisma generate && npx prisma migrate deploy || exit 0",
     "seed": "npx prisma db seed",
+    "seed-ci": "ts-node ./prisma/seed/seed-ci.ts",
     "start": "npm run connectprisma && node build/src/index.js",
     "buildAndStart": "npm run connectprisma && tsc && npm run seed && node build/src/index.js",
     "dev": "nodemon -L -w src -e ts --exec \"npm run buildAndStart\"",
@@ -23,7 +24,7 @@
     "test": "npm run all-tests"
   },
   "prisma": {
-    "seed": "ts-node prisma/seed/seedData.ts"
+    "seed": "./prisma/seed/seed.sh"
   },
   "dependencies": {
     "@prisma/client": "^5.4.1",
